LayoutTests/imported/w3c/web-platform-tests/payment-request/rejects_if_not_active.https-expected.txt


PASS PaymentRequest.show() aborts if the document is not active.
PASS PaymentRequest.show() aborts if the document is active, but not fully active.
PASS PaymentRequest.canMakePayment() rejects if the document is not active.
PASS PaymentRequest.canMakePayment() rejects if the document is active, but not fully active.
